requirements.txt
Mastodon.py
python-dotenv
requests
padatious
html2text
currencies